Combine interface and default impl into one file
